+++ src/test/java/kitchenpos/tmp/application/FakePurgomalumClient.java
@@ -1,6 +1,6 @@
 package kitchenpos.tmp.application;
 
-import kitchenpos.common.PurgomalumClient;
+import kitchenpos.common.infra.PurgomalumClient;
 
 import java.util.Arrays;
 import java.util.List;
